Fix describe_namespace integration spec to use the default namespace (#129)
It can take some time between registering a namespace, and that
namespace being available to describe. Moreover, the test as originally
written can depend on the behavior of a previous run. The description
passed on initial namespace registration will remain on the namespace
indefinitely. If this value is changed in code, subsequent tests
will fail when the description does not match.

Using ruby-samples (or whatever name has been overridden in an
environment variable) ensures the same behavior each run and is not
subject to the possible delay of creating a new namespace since
this namespace must already exist to be able to start the test
workers.