Include remaining changes from upstream #188 (#98)
### Summary

This PR is a follow-up to #97 that includes the changes made to the corresponding upstream PR ([github.com/coinbase/temporal-ruby#188](https://github.com/coinbase/temporal-ruby/pull/188)) after that downstream PR was eagerly merged (with a subset of the changes from upstream). The changes that were not included in the downstream PR when it was merged are: [compare](https://github.com/coinbase/temporal-ruby/pull/188/files/05b6eafeb43ac717c15ae683e6249c4de876ef3d..c6f614325695cc666e066aa218a329c9bf7504f3) (that should be identical to the changes in this PR).

With the merging of the upstream PR, I also updated the [non-upstreamed changed doc](https://paper.dropbox.com/doc/Non-upstreamed-temporal-ruby-changes--Bm39qa99fshz6nbw9RG37pgFAg-AYkwiCfkcoM66adjZMwAZ) to remove the entry on Initial workflow search attributes.

### Motivation
The intention of this PR is to synchronize upstream and downstream, with relation to the changeset applied in the upstream PR.

### Test plan
N/A

### Rollout/monitoring/revert plan
Safe to revert.