do not fail workflow task if completing it errors
It's possible for a workflow task to fail to complete even during normal
operation. For example, if a signal is added to the workflow history
concurrently with a workflow task attempting to complete the workflow
then the `RespondWorkflowTaskCompleted` call will recieve an
`InvalidArgument` error with an `UnhandledCommand` cause.

If this happens then the Ruby SDK would get this error and then try and
fail the workflow task due to how the `rescue` block encompassed the
completion function. This would then lead to the
`RespondWorkflowTaskFailed` call failing because the server doesn't
recognize the task anymore. This is because the task was actually
finalized when the original `RespondWorkflowTaskCompleted` call was made
and so it should not be interacted with anymore.