Add mTLS with Temporal Cloud instructions
The "out of the box" SSL instructions didn't work for me because I used the option to create my own CA certificate and end-end entity certificates.

I added a section to Authentication that makes it easy for people following along with the Temporal Cloud setup. If you create your own CA certificate, including the root certificate in the constructor of GRPC::Core::ChannelCredentials.new leads to auth failures, but the tests of that function ensure that it's okay to pass `nil` as that value, which is needed for people following the Temporal Cloud instructions.

For developer happiness, I think it's also reasonable for this library to include Temporal Cloud-friendly instructions. (not a spokesperson!) :D