Remove deprecated `Temporal.configuration` function
Fixes #143. For a few years there has been a noise `warn` message telling people to not use the function Temporal.configuration, but this function was still used internall in temporal-ruby.

The options were:
a) Remove the logger warn message, as whoever in the past had planned to deprecate access via `Temporal.configuration` never followed up. I don't know why they were trying to remove this method of accessing it.
b) **[CHOSEN OPTION]** Raise the `config` function from `private` to be accessible externally. I don't see why this wouldn't be okay.

I ran the tests per the instructions and I get the same amount of test passes, without the noisy deprecation warnings.

I don't know the full history that led to this being half-deprecated, but this seems like a fine way to clean it up.